+++ smawg/viz.py
@@ -71,6 +71,8 @@ def build_graph(map: Map) -> Graph:
     )
     for i, tile in enumerate(map.tiles):
         node_attrs = {"label": f"{i}. {tile.terrain}"}
+        for symbol in tile.symbols:
+            node_attrs["label"] += f"\n{symbol}"
         if tile.has_a_lost_tribe:
             node_attrs["label"] += "\nLost Tribe"
         if tile.is_at_map_border: